+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - New #199: Add `supportsGuestQueries` field to `User` type.
 - New #199: Add `guestQueryId`, `guestBotCallerUser` and `guestBotCallerChat` fields to `Message` type.
+- New #199: Add `guestMessage` field to `Update` type.
 
 ## 0.20 April 26, 2026